backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=2.8
django-cors-headers>=3.14
gunicorn
django-filter>=23.0
django-grappelli>=2.17.2
Pillow>=9.0
